+++ release1.0/src/Common.h
@@ -238,7 +238,7 @@ string static GetCurrentDir(void) {
 class Common
 {
 public:
-	double static estar(float t);
+	float static estar(float t);
 	static float Qs(float P, float tc);
 	float static Qswater(float P, float tc);
 	float static DepthofSnow(float SWE);
